Restore content to commit 57c0779 (18 memory optimization)
