--- .pre-commit-config.yaml
@@ -1,29 +1,0 @@
-repos:
--   repo: https://github.com/pre-commit/pre-commit-hooks
-    rev: v6.0.0
-    hooks:
-    -   id: trailing-whitespace
-    -   id: end-of-file-fixer
-    -   id: check-yaml
-    -   id: check-added-large-files
-
--   repo: https://github.com/psf/black
-    rev: 25.9.0
-    hooks:
-    -   id: black
-
--   repo: https://github.com/PyCQA/isort
-    rev: 6.1.0
-    hooks:
-    -   id: isort
-        name: isort (python)
-
--   repo: https://github.com/pylint-dev/pylint
-    rev: v3.3.8
-    hooks:
-    -   id: pylint
-        name: pylint
-        entry: poetry run pylint
-        language: system
-        types: [python]
-        args: ["--rcfile=pyproject.toml"]
