--- pyproject.toml
@@ -1,38 +1,0 @@
-[project]
-name = "animate"
-version = "2.0.0"
-description = "An anime recommendation engine."
-authors = [
-    {name = "Asif Sayyed"}
-]
-license = {text = "MIT"}
-readme = "README.md"
-requires-python = ">=3.10"
-
-dependencies = [
-    "pandas (>=2.3.2,<3.0.0)",
-    "scikit-learn (>=1.7.2,<2.0.0)",
-    "nltk (>=3.9.1,<4.0.0)",
-    "streamlit (>=1.50.0,<2.0.0)",
-    "psutil (>=7.1.0,<8.0.0)",
-    "lxml (>=6.0.2,<7.0.0)",
-    "beautifulsoup4 (>=4.14.2,<5.0.0)",
-    "requests (>=2.32.5,<3.0.0)",
-    "tqdm (>=4.67.1,<5.0.0)",
-]
-
-
-[build-system]
-requires = ["poetry-core>=2.0.0,<3.0.0"]
-build-backend = "poetry.core.masonry.api"
-
-[tool.poetry.group.dev.dependencies]
-ipykernel = "^6.30.1"
-jupyter = "^1.1.1"
-pre-commit = "^4.3.0"
-isort = "^7.0.0"
-black = "^25.9.0"
-pylint = "^4.0.1"
-
-[tool.pylint.format]
-max-line-length = 120
