+++ requirements.txt
@@ -1,0 +1,110 @@
+altair==5.3.0
+annotated-types==0.7.0
+asttokens==2.4.1
+asyncio==3.4.3
+attrs==23.2.0
+beautifulsoup4==4.12.3
+blinker==1.8.2
+blis==0.7.11
+bs4==0.0.2
+cachetools==5.4.0
+catalogue==2.0.10
+certifi==2024.7.4
+charset-normalizer==3.3.2
+click==8.1.7
+cloudpathlib==0.18.1
+colorama==0.4.6
+comm==0.2.2
+confection==0.1.5
+contourpy==1.2.1
+cycler==0.12.1
+cymem==2.0.8
+debugpy==1.8.2
+decorator==5.1.1
+en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
+en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
+executing==2.0.1
+fonttools==4.53.1
+gitdb==4.0.11
+GitPython==3.1.43
+greenlet==3.0.3
+idna==3.7
+ipykernel==6.29.5
+ipython==8.26.0
+jedi==0.19.1
+Jinja2==3.1.4
+joblib==1.4.2
+jsonschema==4.23.0
+jsonschema-specifications==2023.12.1
+jupyter_client==8.6.2
+jupyter_core==5.7.2
+kiwisolver==1.4.5
+langcodes==3.4.0
+language_data==1.2.0
+marisa-trie==1.2.0
+markdown-it-py==3.0.0
+MarkupSafe==2.1.5
+matplotlib==3.9.1
+matplotlib-inline==0.1.7
+mdurl==0.1.2
+murmurhash==1.0.10
+nest-asyncio==1.6.0
+nltk==3.9.1
+numpy==1.26.4
+packaging==24.1
+pandas==2.2.2
+parso==0.8.4
+pillow==10.4.0
+platformdirs==4.2.2
+playwright==1.45.1
+preshed==3.0.9
+prompt_toolkit==3.0.47
+protobuf==5.27.2
+psutil==6.0.0
+pure_eval==0.2.3
+pyarrow==17.0.0
+pydantic==2.8.2
+pydantic_core==2.20.1
+pydeck==0.9.1
+pyee==11.1.0
+Pygments==2.18.0
+pyparsing==3.1.2
+python-dateutil==2.9.0.post0
+pytz==2024.1
+pyzmq==26.0.3
+referencing==0.35.1
+regex==2024.7.24
+requests==2.32.3
+rich==13.7.1
+rpds-py==0.19.1
+scikit-learn==1.5.1
+scipy==1.14.0
+seaborn==0.13.2
+shellingham==1.5.4
+six==1.16.0
+smart-open==7.0.4
+smmap==5.0.1
+soupsieve==2.5
+spacy==3.7.6
+spacy-legacy==3.0.12
+spacy-loggers==1.0.5
+srsly==2.4.8
+stack-data==0.6.3
+streamlit==1.36.0
+tenacity==8.5.0
+thinc==8.2.5
+threadpoolctl==3.5.0
+toml==0.10.2
+toolz==0.12.1
+tornado==6.4.1
+tqdm==4.66.4
+traitlets==5.14.3
+typer==0.12.5
+typing_extensions==4.12.2
+tzdata==2024.1
+urllib3==2.2.2
+wasabi==1.1.3
+watchdog==4.0.1
+wcwidth==0.2.13
+weasel==0.4.1
+wrapt==1.16.0
